Get rid of hook
